minor refactoring before stating v2 changes.
