+++ middlewares/osio/osio_basic_test.go
@@ -1,0 +1,236 @@
+package middlewares
+
+import (
+	"fmt"
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"strings"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+type testMiddlewareCtx struct {
+	tables          []testMiddlewareData
+	authCallCount   int
+	tenantCallCount int
+	currInd         int
+}
+
+type testMiddlewareData struct {
+	inputPath      string
+	inputAuth      string
+	expectedTarget string
+	expectedAuth   string
+	expectedPath   string
+}
+
+var mwCtx = testMiddlewareCtx{tables: []testMiddlewareData{
+	{
+		"/",
+		"Bearer 1000",
+		"http://api.cluster1.com",
+		"Bearer 1001",
+		"/",
+	},
+	{
+		"/api",
+		"Bearer 1000",
+		"http://api.cluster1.com",
+		"Bearer 1001",
+		"/api",
+	},
+	{
+		"/api/api",
+		"Bearer 1000",
+		"http://api.cluster1.com",
+		"Bearer 1001",
+		"/api",
+	},
+	{
+		"/api/anything",
+		"Bearer 1000",
+		"http://api.cluster1.com",
+		"Bearer 1001",
+		"/api/anything",
+	},
+	{
+		"/api/api/anything",
+		"Bearer 1000",
+		"http://api.cluster1.com",
+		"Bearer 1001",
+		"/api/anything",
+	},
+	{
+		"/oapi/anything",
+		"Bearer 1000",
+		"http://api.cluster1.com",
+		"Bearer 1001",
+		"/oapi/anything",
+	},
+	{
+		"/api/oapi/anything",
+		"Bearer 1000",
+		"http://api.cluster1.com",
+		"Bearer 1001",
+		"/oapi/anything",
+	},
+	{
+		"/metrics",
+		"Bearer 1000",
+		"http://metrics.cluster1.com",
+		"Bearer 1001",
+		"/",
+	},
+	{
+		"/metrics/anything",
+		"Bearer 1000",
+		"http://metrics.cluster1.com",
+		"Bearer 1001",
+		"/anything",
+	},
+	{
+		"/restall",
+		"Bearer 1000",
+		"http://api.cluster1.com",
+		"Bearer 1001",
+		"/restall",
+	},
+	{
+		"/api",
+		"Bearer 2000",
+		"http://api.cluster1.com",
+		"Bearer 2001",
+		"/api",
+	},
+	{
+		"/metrics",
+		"Bearer 2000",
+		"http://metrics.cluster1.com",
+		"Bearer 2001",
+		"/",
+	},
+}}
+
+func TestMiddleware(t *testing.T) {
+	os.Setenv("AUTH_TOKEN_KEY", "foo")
+
+	tenantServer := mwCtx.createServer(mwCtx.serveTenantRequest)
+	defer tenantServer.Close()
+	authServer := mwCtx.createServer(mwCtx.serverAuthRequest)
+	defer authServer.Close()
+
+	tenantURL := "http://" + tenantServer.Listener.Addr().String() + "/"
+	authURL := "http://" + authServer.Listener.Addr().String() + "/"
+	srvAccID := "sa1"
+	srvAccSecret := "secret"
+
+	osio := NewOSIOAuth(tenantURL, authURL, srvAccID, srvAccSecret)
+	osio.RequestTokenType = mwCtx.testTokenTypeLocator
+	osioServer := mwCtx.createServer(mwCtx.serverOSIORequest(osio))
+	osioURL := osioServer.Listener.Addr().String()
+
+	for ind, table := range mwCtx.tables {
+		mwCtx.currInd = ind
+
+		currReqPath := table.inputPath
+		currAuth := table.inputAuth
+
+		req, _ := http.NewRequest("GET", "http://"+osioURL+currReqPath, nil)
+		req.Header.Set("Authorization", currAuth)
+		res, _ := http.DefaultClient.Do(req)
+		assert.NotNil(t, res)
+		assert.Equal(t, http.StatusOK, res.StatusCode)
+		errMsg := res.Header.Get("err")
+		assert.Empty(t, errMsg, errMsg)
+	}
+
+	// validate cache is used
+	expectedTenantCalls := 2
+	expectedAuthCalls := 2
+	assert.Equal(t, expectedTenantCalls, mwCtx.tenantCallCount, "Number of time Tenant server called was incorrect, want:%d, got:%d", expectedTenantCalls, mwCtx.tenantCallCount)
+	assert.Equal(t, expectedAuthCalls, mwCtx.authCallCount, "Number of time Auth server called was incorrect, want:%d, got:%d", expectedAuthCalls, mwCtx.authCallCount)
+}
+
+func (t testMiddlewareCtx) createServer(handle func(http.ResponseWriter, *http.Request)) *httptest.Server {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/", handle)
+	return httptest.NewServer(mux)
+}
+
+func (t testMiddlewareCtx) serveTenantRequest(rw http.ResponseWriter, req *http.Request) {
+	mwCtx.tenantCallCount++
+	user := ""
+	if strings.HasSuffix(req.Header.Get(Authorization), "1000") {
+		user = "1000"
+	} else if strings.HasSuffix(req.Header.Get(Authorization), "2000") {
+		user = "2000"
+	}
+	res := fmt.Sprintf(`{
+		"data": {
+			"attributes": {
+				"namespaces": [
+					{
+						"name": "myuser%s-preview-stage",
+						"type": "user",
+						"cluster-metrics-url": "http://metrics.cluster1.com",
+						"cluster-url": "http://api.cluster1.com"
+					}
+				]
+			}
+		}
+	}`, user)
+	rw.Write([]byte(res))
+}
+
+func (t testMiddlewareCtx) serverAuthRequest(rw http.ResponseWriter, req *http.Request) {
+	mwCtx.authCallCount++
+	token := "1001"
+	if strings.HasSuffix(req.Header.Get(Authorization), "1000") {
+		token = "1001"
+	} else if strings.HasSuffix(req.Header.Get(Authorization), "2000") {
+		token = "2001"
+	}
+	res := fmt.Sprintf(`{"token_type":"bearer", "scope":"user","access_token":"%s"}`, token)
+	rw.Write([]byte(res))
+}
+
+func (t testMiddlewareCtx) serverOSIORequest(osio *OSIOAuth) func(http.ResponseWriter, *http.Request) {
+	return func(rw http.ResponseWriter, req *http.Request) {
+		osio.ServeHTTP(rw, req, mwCtx.varifyHandler)
+	}
+}
+
+func (t testMiddlewareCtx) varifyHandler(rw http.ResponseWriter, req *http.Request) {
+	expectedTarget := mwCtx.tables[mwCtx.currInd].expectedTarget
+	actualTarget := req.Header.Get("Target")
+	if expectedTarget != actualTarget {
+		rw.Header().Set("err", fmt.Sprintf("Target was incorrect, want:%s, got:%s", expectedTarget, actualTarget))
+		return
+	}
+
+	expectedAuth := mwCtx.tables[mwCtx.currInd].expectedAuth
+	actualAuth := req.Header.Get(Authorization)
+	if expectedAuth != actualAuth {
+		rw.Header().Set("err", fmt.Sprintf("Authorization was incorrect, want:%s, got:%s", expectedAuth, actualAuth))
+		return
+	}
+
+	expectedPath := mwCtx.tables[mwCtx.currInd].expectedPath
+	actualPath := req.URL.Path
+	if expectedPath != actualPath {
+		rw.Header().Set("err", fmt.Sprintf("Path was incorrect, want:%s, got:%s", expectedPath, actualPath))
+		return
+	}
+
+	actualPath = req.RequestURI
+	if expectedPath != actualPath {
+		rw.Header().Set("err", fmt.Sprintf("Path was incorrect, want:%s, got:%s", expectedPath, actualPath))
+		return
+	}
+}
+
+func (t testMiddlewareCtx) testTokenTypeLocator(token string) (TokenType, error) {
+	return UserToken, nil
+}
